+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## master
+
+* [improvement] user can answer 'y' to 'yes/no' questions
+
 ## 0.2.27 / 2013-06-11
 
 * [feature] user is able to download logs for a given day by `shelly logs get [DATE]`